requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub
groq
langchain-groq
beautifulsoup4
sentence-transformers

#-e .